worklady/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.61.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
idna==3.7
jwt==1.3.1
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
